+++ packages/transaction-controller/src/utils/batch.ts
@@ -246,13 +246,14 @@ async function getNestedTransactionMeta(
   ethQuery: EthQuery,
 ): Promise<NestedTransactionMetadata> {
   const { from } = request;
-  const { params } = singleRequest;
+  const { params, type: requestedType } = singleRequest;
 
-  const { type } = await determineTransactionType(
+  const { type: determinedType } = await determineTransactionType(
     { from, ...params },
     ethQuery,
   );
 
+  const type = requestedType ?? determinedType;
   return {
     ...params,
     type,
@@ -550,7 +551,7 @@ async function processTransactionWithHook(
   request: AddTransactionBatchRequest,
   txBatchMeta?: TransactionBatchMeta,
 ) {
-  const { existingTransaction, params } = nestedTransaction;
+  const { existingTransaction, params, type } = nestedTransaction;
 
   const {
     addTransaction,
@@ -606,6 +607,7 @@ async function processTransactionWithHook(
       networkClientId,
       publishHook,
       requireApproval: false,
+      type,
     },
   );
 
@@ -626,11 +628,16 @@ async function processTransactionWithHook(
     value,
   };
 
-  log('Processed new transaction with hook', { id, params: newParams });
+  log('Processed new transaction with hook', {
+    id,
+    params: newParams,
+    type,
+  });
 
   return {
     id,
     params: newParams,
+    type,
   };
 }